Undo PR #1561
In https://github.com/artefactual/archivematica/pull/1561 we removed
the creation of the base directory for a Dspace transfer from the
`fileToFolder_v1.0` client script because the MCPServer created the
directory in advance.

Since we're removing that capability from the MCPServer in this PR, we
need to restore the old logic in the client script back.

+++ src/MCPClient/lib/clientScripts/file_to_folder.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python
 
 import os
+import sys
 
 import django
 
@@ -32,6 +33,17 @@ def main(job, transfer_path, transfer_uuid, shared_path):
     # If file, move into directory and update transfer
     dirpath = os.path.splitext(transfer_path)[0]
     basename = os.path.basename(transfer_path)
+    if os.path.exists(dirpath):
+        job.pyprint(
+            "Cannot move file",
+            transfer_path,
+            "to folder",
+            dirpath,
+            "because it already exists",
+            file=sys.stderr,
+        )
+        return 1
+    os.mkdir(dirpath)
     new_path = os.path.join(dirpath, basename)
     job.pyprint("Moving", transfer_path, "to", new_path)
     os.rename(transfer_path, new_path)
